Introducing useTbSelection
